Removed calc component, moved the class to singleton.py
